change input numeber type int into float

+++ testPT.c
@@ -2,28 +2,28 @@
 #include"pt.h"
 
 static struct pt inputPT,calculatePT;
-static unsigned int num1,num2,result;
+static float num1,num2,result;
 static char op;
 
 static PT_THREAD(input(struct pt *pt)){
 	PT_BEGIN(pt);
 	printf("Input first number!\n");
-	scanf("%d",&num1);
+	scanf("%f",&num1);
 	printf("Input a operation!\n");
 	getchar();
         scanf("%c",&op);
 	printf("Input another number!\n");
-        scanf("%d",&num2);
+        scanf("%f",&num2);
 	PT_END(pt);
 }
 
 static PT_THREAD(calculate(struct pt *pt)){
 	PT_BEGIN(pt);
 	printf("The result is ");
-	if( '+' == op){	result=num1+num2;	printf("%d + %d = %d\n",num1,num2,result);	}
-	if( '-' == op){	result=num1-num2;	printf("%d - %d = %d\n",num1,num2,result);	}
-	if( '*' == op){	result=num1*num2;	printf("%d * %d = %d\n",num1,num2,result);	}
-	if( '/' == op){	result=num1/num2;	printf("%d / %d = %d\n",num1,num2,result);	}
+	if( '+' == op){	result=num1+num2;	printf("%f + %f = %f\n",num1,num2,result);	}
+	if( '-' == op){	result=num1-num2;	printf("%f - %f = %f\n",num1,num2,result);	}
+	if( '*' == op){	result=num1*num2;	printf("%f * %f = %f\n",num1,num2,result);	}
+	if( '/' == op){	result=num1/num2;	printf("%f / %f = %f\n",num1,num2,result);	}
 	PT_END(pt);
 }
